tap-94 Load affect lexicon on startup
