+++ src/main/scala/tap/nlp/NlpInitialiserModule.scala
@@ -19,6 +19,7 @@ package tap.nlp
 import com.google.inject.AbstractModule
 import play.api.Logger
 import play.api.libs.concurrent.AkkaGuiceSupport
+import tap.analysis.affectlexicon.AffectLexiconActor
 import tap.nlp.factorie.LanguageToolActor
 
 /**
@@ -30,6 +31,7 @@ class NlpInitialiserModule extends AbstractModule with AkkaGuiceSupport {
   def configure():Unit = {
     Logger.info("Binding LanguageToolActor")
     bindActor[LanguageToolActor]("languagetool")
+    bindActor[AffectLexiconActor]("affectlexicon")
   }
 
 }
